frontend and tax records

+++ front/components/Tables/OverviewTransactionInputs.vue
@@ -1,17 +1,17 @@
 <template>
     <div>
-        <b-button variant="outline-danger" :disabled="buttonDisabled" @click="removeAll">Delete</b-button>
+        <b-button v-if="transactionInputs.length > 0" variant="outline-danger" class="align-right" :disabled="buttonDisabled" @click="removeAll">Delete All</b-button>
         <b-tabs pills content-class="mt-3">
             <b-tab title="Total" active>
                 <b-card
-                    v-if="transaction_inputs.length === 0"
+                    v-if="transactionInputs.length === 0"
                     sub-title="An Overview Of Transactions Will Appear Here Once You Start Uploading Data"
                     class="text-center py-5"
                 ></b-card>
                 <table-transaction-inputs v-else />
 
             </b-tab>
-            <b-tab v-for="account in seller_firm.accounts" :key="account.public_id" :title="account.channel_code" lazy>
+            <b-tab v-for="account in sellerFirm.accounts" :key="account.public_id" :title="account.channel_code" lazy>
                 <lazy-table-transaction-inputs :channelCode="account.channel_code" />
                 <!-- {{ account.channel_code }} -->
             </b-tab>
@@ -38,22 +38,26 @@
 
         computed: {
             ...mapState({
-                seller_firm: state => state.seller_firm.seller_firm,
-                transaction_inputs: state => state.transaction_input.transaction_inputs
+                sellerFirm: state => state.seller_firm.seller_firm,
+                transactionInputs: state => state.transaction_input.transaction_inputs
             }),
         },
 
         methods: {
             async removeAll() {
                 this.buttonDisabled = true;
-                try {
-                    await this.$store.dispatch("transaction_input/delete_all");
-                    await this.$store.dispatch("seller_firm/get_by_public_id", this.$route.params.public_id);
-                    this.buttonDisabled = false;
-                } catch (error) {
-                    this.$toast.error(error, { duration: 5000 });
-                    this.buttonDisabled = false;
-                    return [];
+                if (this.transactionInputs.length > 0) {
+                    try {
+
+                        await this.$store.dispatch("transaction_input/delete_all");
+                        await this.$store.dispatch("seller_firm/get_by_public_id", this.$route.params.public_id);
+                        this.$router.push(`/tax/clients/${this.$route.params.public_id}`)
+
+                    } catch (error) {
+                        this.$toast.error(error, { duration: 5000 });
+                        this.buttonDisabled = false;
+                        return [];
+                    }
                 }
             },
         },
